zod with sign up working
