+++ package.json
@@ -21,7 +21,8 @@
     "pg": "^8.11.5",
     "react": "^19.0.0-beta-4508873393-20240430",
     "react-dom": "^19.0.0-beta-4508873393-20240430",
-    "react-hook-form": "^7.51.3"
+    "react-hook-form": "^7.51.3",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
