+++ src/app/actions.ts
@@ -4,66 +4,108 @@ import { cookies } from "next/headers";
 import jwt from "jsonwebtoken";
 import Database from "./db.ts";
 import { redirect } from "next/navigation";
+import { z } from 'zod'
 
-export async function signUp(
-  email: string,
-  first_name: string,
-  last_name: string,
-  password: string,
-  confirm_password: string
-) {
-  let userSignedUp = false;
 
-  const passwordHasher = new PasswordHasher();
-  const hashedPassword = await passwordHasher.hash(password);
-  const db = new Database();
+const schemaSignup = z.object({
+  email: z.string({
+    invalid_type_error: 'Invalid Email',
+  }),
+  first_name: z.string({
+    invalid_type_error: 'Invalid First Name',
+  }),
+  last_name: z.string({
+    invalid_type_error: 'Invalid Last Name',
+  }),
+  password: z.string({
+    invalid_type_error: 'Invalid Password',
+  }),
+  confirm_password: z.string({
+    invalid_type_error: 'Invalid Confirm Password',
+  }),
+})
+
+const schemaLogin = z.object({
+  email: z.string({
+    invalid_type_error: 'Invalid Email',
+  }),
+  password: z.string({
+    invalid_type_error: 'Invalid Password',
+  }),
+})
+
+export async function signUp(prevState: any, formData: FormData) {
+
+  const validatedFields = schemaSignup.safeParse({
+    email: formData.get('email'),
+    first_name: formData.get('first_name'),
+    last_name: formData.get('last_name'),
+    password: formData.get('password'),
+    confirm_password: formData.get('confirm_password'),
+  })
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+    }
+  }
+
+  const email = formData.get("email") as string;
+  const first_name = formData.get("first_name") as string;
+  const last_name = formData.get("last_name") as string;
+  const password = formData.get("password") as string;
+  const confirm_password = formData.get("confirm_password") as string;
+
   try {
     if (password !== confirm_password) {
       throw new Error("Passwords do not match");
     }
 
-    if (await db.register(email, first_name, last_name, hashedPassword)) {
-      userSignedUp = true;
-    } else throw new Error("Database rejected sign up!");
+    const passwordHasher = new PasswordHasher();
+    const hashedPassword = await passwordHasher.hash(password);
+    const db = new Database();
+
+    if (!await db.register(email, first_name, last_name, hashedPassword))
+      throw new Error("Database rejected sign up!");
+
   } catch (error) {
     console.error(
       "Error registering user: ",
       error instanceof Error ? error.message : error
     );
-   
+    return { message: "User not signed up Error: " + (error instanceof Error ? error.message : error) };
   }
-  if (userSignedUp) {
-    applyCookie(email);
-    //redirect(`/dashboard/${email}`);
-    return true;
-  }
-  return false;
+
+  applyCookie(email);
+  redirect(`/dashboard/${email}`);
 }
+
 export async function login(prevState: any, formData: FormData) {
-  let userAuthed = false;
+
+  const validatedFields = schemaSignup.safeParse({
+    email: formData.get('email'),
+    password: formData.get('password'),
+  })
+
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+    }
+  }
   const email = formData.get("email") as string;
   const password = formData.get("password") as string;
   try {
 
-    if (await verifyPassword(email, password)) {
-      userAuthed = true;
-    } else {
+    if (!await verifyPassword(email, password))
       throw new Error("Authentication failed");
-    }
   } catch (error) {
     console.error(
       "Error logging in user: ",
       error instanceof Error ? error.message : error
     );
-    //return {message: "Error logging in user "};
+    return { message: "User not authenticated Error: " + (error instanceof Error ? error.message : error) };
   }
-  // error instanceof Error ? error.message : error
-  if (userAuthed) {
-    applyCookie(email);
-    redirect(`/dashboard/${email}`);
-  } else {
-    return { message: "Invalid credentials" };
-  }
+  applyCookie(email);
+  redirect(`/dashboard/${email}`);
 }
 
 export async function verifyToken(email: string) {
