+++ src/app/login/page.module.css
@@ -1,0 +1,88 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+  width:100%;
+  background-image: linear-gradient(to bottom, rgba(81, 66, 212, 0.856), rgb(96, 152, 202) 100%);
+
+}
+.container {  
+  background-image: linear-gradient(to bottom, rgba(130, 139, 194, 0.801), rgba(96, 174, 202, 0.795) 100%);  
+  width: 70%;
+  min-height:600px;
+  border-radius: 10px;  
+  box-shadow: 0 0 15px 0 rgba(0, 0, 0, 0.1);
+  
+}
+.left {
+  width: 50%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  float: left;
+  padding-top:30px;
+  text-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+
+}
+.left>h4{
+  margin-top: 70px;
+}
+.form{
+  width: 80%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding-top: 40px;
+  gap: 20px;
+  
+}
+
+.right {
+  width: 50%;
+  height: 600px;
+  display: flex;
+  flex-direction: column;
+  /* justify-content: center; */
+  align-items: center;
+  float: right;
+  padding-top:30px;
+  background-color: rgb(84, 113, 218);
+  border-top-right-radius: 10px;
+  border-bottom-right-radius: 10px;
+  text-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+  gap:100px;
+}
+/* .right>h4{
+  margin-top: 70px;
+}
+
+.right>Button{
+  margin-top: 100px;
+} */
+@media screen and (max-width:1250px) {
+  .container {
+    width: 90%;
+  }
+  
+}
+
+@media screen and (max-width:800px) {
+  .container {
+    width: 100%;
+  }
+  .left {
+    width: 100%;
+    margin-bottom:40px;
+  }
+  .right {
+    width: 100%;
+    height:400px;
+    
+  }
+}
